+++ docs/CHANGELOG.md
@@ -10,7 +10,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ...
 
-## 1.13.0 - UNRELEASED
+<!-- unreleasedstop -->
+
+## 1.13.0 - 2023-03-16
 
 ### Added
 
@@ -29,8 +31,6 @@ of any values provided via configuration files.
 ### Removed
 
 - Terra is no longer supported
-
-<!-- unreleasedstop -->
 
 ## 1.12.0 - 2023-02-15